requirements.txt
Flask
PyMySQL
requests
python-dotenv
flask-login
werkzeug

